LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/security.drawImage.canvas-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
security.drawImage.canvas
drawImage of unclean canvas makes the canvas origin-unclean
Actual output:

PASS drawImage of unclean canvas makes the canvas origin-unclean
